Fix compilation failure on old gcc, avoid compiler warnings
std::underlying_type isn't allowed to be used unless you already know
the type is an enum (until C++20), and causes a compilation failure if
used (even after the boolean check) on gcc 8 and below.

The bind_blob and bind_text taking over ownership might be useful, but
currently aren't used so just comment them out so if someone develops a
need they are ready to go.