Make blink signature parsing single-pass
This signature parsing code was bizarrely failing to parse what appears
to be perfectly valid data, but only on one Debian buster system for
unknown reasons, and not reproducible in a small test case with the
exact data it received.

Rewrote it to use on-the-fly data deserialization (which is more
efficient anyway by avoiding heap memory allocation for
deserialization).