Add grace to required difficulty - fix 526484
At height 526483 a mined block arrived that had *slightly* too little
proof of work, yet difficulties on the network had apparently diverged
enough that half the network accepted it and the other half didn't --
nodes with the correct difficulty refuse it, so a recalculate difficulty
cannot help.

This hacks around the problem from 526483 until the next hard fork by
adding a grace to the accepted block difficulty - we'll accept a block
from the network if it has 99.8% of the calculated difficulty value.

It also defaults --recalculate-difficulty to 1 so that lokid does a full
difficulty rescan at startup.  (This seems to only take a second or two
so seems like a good thing to enable by default).