Fix tx serialization tests
Some of the invalid bulletproof tests were relying on the broken
behaviour of tx serialization failure still producing a tx hash based on
a partially filled serialized blob.  The code now checks for
serialization failures, and throws when trying to get a tx hash from a
tx that fails to serialize.  (This isn't new behaviour: it's just that
some serialization failures didn't actually signal a failure properly).

Fix up the tests to use a random tx hash if such a serialization
exception occurs, so that we can still add the invalid txes to blocks
for the purposes of testing block/tx failures.