Add drone CI config
Adds various builds and does a basic offline startup test on all, plus
runs the full test suite on a debian and a mac build.

This has the ability to do an all-static build and to build debs, and
upload both to builds.lokinet.dev, but it is not yet enabled and will
likely take some additional work.