Don't use external project for no reason
Using an external project to build a subdirectory is gross, and moreover
it breaks if you are trying to use a custom compiler and uses the wrong
one (or just fails if a `c++` binary doesn't exist).

Since the builds appear to run just fine without this, just include it
via add_subdirectory instead.