Fix crash from race condition in miner::stop()
miner::stop() destroys threads without joining them, which can call
`std::terminate()` and abort the program if any of the threads are still
active (which happens to me consistently under --regtest).

This fixes the code to properly join threads rather than using a custom
racey hack with a sleep loop.