Fix interactive prepare_registration under --regtest
The interactive mode handler for `prepare_registration` didn't properly
detect FAKECHAIN mode and so would show mainnet values in the prompts;
this fixes it to notice FAKECHAIN mode properly.

I did this by just checking the `.nettype` string rather than adding a
new field since fakechain is a dev tool that didn't seem to warrant a
new field in the RPC struct.