LNS: multi-row update simplification
This replaces the way to handle update the remove the need to use the
registration height and prev_txid fields: instead we simply store a new
row copied from the old row with changes applied as needed.

Selecting values is then a matter of just selecting the row with the
highest update_height.

The registration_height isn't needed anymore, either, since we now have
an expiration_height (i.e. we don't to know when it was registered); now
we can just use the update_height.

txid is still present and effectively still serves as a nonce to prevent
duplicate/conflicting updates: the new TX still needs to reference the
previous txid, but we don't need to also store the prev_txid because we
don't need to walk back the chain anymore.

Rollbacks then become *significantly* easier: we just delete anything
with an update_height >= the new blockchain height: the new
"highest-update_height" row will now have the proper values.

Not included in this commit but still to come is pruning the database to
remove heights that are no longer needed (i.e. have been superceded by a
more recent update row more than 5000 blocks ago).